feat: lift ProjectSelector and ProjectSelectorNavigator to ui-core
Move ProjectSelector, ProjectSelectorNavigator, ProjectNavigatorLink,
and ProjectIconWithSize components from frontend/src/concepts/projects/
to @odh-dashboard/ui-core/components/projectSelector.

Also moves IconSize enum and Namespace type to ui-core/src/types.ts.

Updates 14 package consumers to import from ui-core directly:
- agent-ops (2 files)
- automl (1 file)
- autorag (1 file)
- eval-hub (1 file)
- gen-ai (1 file)
- maas (1 file)
- mlflow-embedded (2 files)
- model-registry (1 file)
- model-serving (3 files)
- model-training (1 file)

Adds react-router-dom as a peer dependency of ui-core.
Adds @odh-dashboard/ui-core dependency to gen-ai package.
Frontend backward compatibility preserved via re-export shims.

Ref: RHOAIENG-65097
