+++ frontend/src/concepts/projects/ProjectIconWithSize.tsx
@@ -1,13 +1,2 @@
-import React from 'react';
-import { ProjectIcon } from '#~/images/icons';
-import { IconSize } from '#~/types';
-
-export const ProjectIconWithSize: React.FC<{ size: IconSize }> = ({ size }) => (
-  <ProjectIcon
-    alt=""
-    style={{
-      width: `var(--pf-t--global--icon--size--font--${size})`,
-      height: `var(--pf-t--global--icon--size--font--${size})`,
-    }}
-  />
-);
+// eslint-disable-next-line @odh-dashboard/no-restricted-imports -- re-exporting shared component
+export { ProjectIconWithSize } from '@odh-dashboard/ui-core/components/projectSelector/ProjectIconWithSize';
