+++ packages/agent-ops/frontend/src/app/components/AgentOpsProjectSelector.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import ProjectSelector from '@odh-dashboard/internal/concepts/projects/ProjectSelector';
+import ProjectSelector from '@odh-dashboard/ui-core/components/projectSelector/ProjectSelector';
 import { FormHelperText, HelperText, HelperTextItem } from '@patternfly/react-core';
 import { useNavigate } from 'react-router-dom';
 import {
