+++ packages/agent-ops/frontend/src/app/deployWizard/steps/ImageSelectionStep.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import ProjectSelector from '@odh-dashboard/internal/concepts/projects/ProjectSelector';
+import ProjectSelector from '@odh-dashboard/ui-core/components/projectSelector/ProjectSelector';
 import {
   Form,
   FormGroup,
