+++ packages/model-serving/src/components/deploymentWizard/fields/ProjectSection.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import { FormGroup, HelperText, HelperTextItem, TextInput } from '@patternfly/react-core';
-import ProjectSelector from '@odh-dashboard/internal/concepts/projects/ProjectSelector';
+import ProjectSelector from '@odh-dashboard/ui-core/components/projectSelector/ProjectSelector';
 import { ProjectsContext } from '@odh-dashboard/ui-core/context/ProjectsContext';
 import { byName } from '@odh-dashboard/k8s-core';
 import { ODH_PRODUCT_NAME } from '@odh-dashboard/ui-core/utilities';
