+++ packages/model-serving/src/components/deploymentWizard/steps/PreconfigureDeploymentStep.tsx
@@ -9,7 +9,7 @@ import {
 } from '@patternfly/react-core';
 import { OutlinedQuestionCircleIcon } from '@patternfly/react-icons';
 import { ODH_PRODUCT_NAME } from '@odh-dashboard/ui-core/utilities';
-import ProjectSelector from '@odh-dashboard/internal/concepts/projects/ProjectSelector';
+import ProjectSelector from '@odh-dashboard/ui-core/components/projectSelector/ProjectSelector';
 import { UseModelDeploymentWizardState } from '../useDeploymentWizard';
 
 type PreconfigureDeploymentStepProps = {
