+++ packages/ui-core/package.json
@@ -59,7 +59,13 @@
     "./contexts/ThemeContext": "./src/contexts/ThemeContext.tsx",
     "./contexts/AnalyticsContext": "./src/contexts/AnalyticsContext.tsx",
     "./contexts/NotificationContext": "./src/contexts/NotificationContext.tsx",
-    "./contexts/NotificationWatcherContext": "./src/contexts/NotificationWatcherContext.tsx"
+    "./contexts/NotificationWatcherContext": "./src/contexts/NotificationWatcherContext.tsx",
+    "./types": "./src/types.ts",
+    "./components/projectSelector": "./src/components/projectSelector/index.ts",
+    "./components/projectSelector/ProjectSelector": "./src/components/projectSelector/ProjectSelector.tsx",
+    "./components/projectSelector/ProjectSelectorNavigator": "./src/components/projectSelector/ProjectSelectorNavigator.tsx",
+    "./components/projectSelector/ProjectNavigatorLink": "./src/components/projectSelector/ProjectNavigatorLink.tsx",
+    "./components/projectSelector/ProjectIconWithSize": "./src/components/projectSelector/ProjectIconWithSize.tsx"
   },
   "scripts": {
     "lint": "eslint .",
@@ -87,6 +93,7 @@
     "@patternfly/react-styles": "^6",
     "@patternfly/react-table": "^6",
     "react": "^18",
+    "react-router-dom": "^6",
     "zod": "^3"
   }
 }
