+++ packages/ui-core/src/components/projectSelector/index.ts
@@ -1,0 +1,5 @@
+/* eslint-disable no-barrel-files/no-barrel-files */
+export { default as ProjectSelector } from './ProjectSelector';
+export { default as ProjectSelectorNavigator } from './ProjectSelectorNavigator';
+export { default as ProjectNavigatorLink } from './ProjectNavigatorLink';
+export { ProjectIconWithSize } from './ProjectIconWithSize';
